Add files via upload
A simple 'to do'-list that allows the users to add items to a list of their own. Each item added by the user is stored into an array.

The project has to run locally on the device or on a server, in order to be operational. 

The following npm-modules are required and used in this project:
- express
- bodyParser